+++ src/api/handler/auth/index.ts
@@ -1,2 +1,2 @@
-export * from "./|api|signin";
-export * from "./|api|signout";
+export * from "./signin";
+export * from "./signout";
