+++ src/api/handler/auth/signout.ts
